requirements.txt
langchain
langchain_community
langchain_core
langchain_openai
python-dotenv
streamlit
chromadb
beautifulsoup4
google-generativeai
langchain_chroma